refactor all ID's out and replace with symbol; minor fixes

+++ src/main/java/org/sci/finalproj/service/AssetService.java
@@ -32,10 +32,9 @@ public class AssetService {
         return assetsList;
     }
 
-    public void buyCryptoAsset(double amount, String cryptoCoinSymbol, Long defaultWalletCurrencyId, Long userId) {
-        Asset defaultAsset = assetRepo.findByCoinId(defaultWalletCurrencyId);
-
-        if (defaultAsset.getCoinAmount() >= amount) {
+    public void buyCryptoAsset(double amount, String cryptoCoinSymbol, String defaultCurrencySymbol, Long userId) {
+        Asset defaultAsset = assetRepo.findByCoinSymbol(defaultCurrencySymbol);
+        if (defaultAsset != null && defaultAsset.getCoinAmount() >= amount) {
             if (defaultAsset.getCoinAmount() == amount) {
                 assetRepo.deleteById(defaultAsset.getAssetId());
             } else {
@@ -45,13 +44,11 @@ public class AssetService {
             System.out.println("Please input card details: to be implemented; keep going for now");
         }
 
-        CryptoCoin cryptoCoin = cryptoCoinRepo.findByCryptoCoinSymbol(cryptoCoinSymbol);
-        long cryptoId = cryptoCoin.getCryptoCoinId();
-        Asset existentCryptoAsset = assetRepo.findByCoinId(cryptoId);
+        Asset existentCryptoAsset = assetRepo.findByCoinSymbol(cryptoCoinSymbol);
         // add or update asset :
         if (existentCryptoAsset == null) {
             //add
-            assetRepo.save(new Asset(userId, cryptoId, cryptoCoinSymbol, amount));
+            assetRepo.save(new Asset(userId, cryptoCoinSymbol, amount));
         } else {
             //update
             double existentAmount = existentCryptoAsset.getCoinAmount();
@@ -59,18 +56,14 @@ public class AssetService {
         }
         // add transaction to DB :
         Date currentSqlDate = new Date(System.currentTimeMillis());
-        transactionRepo.save(new Transaction(userId, defaultAsset.getCoinSymbol(), cryptoCoinSymbol, TransactionType.FICR, amount, currentSqlDate));
+        transactionRepo.save(new Transaction(userId, defaultCurrencySymbol, cryptoCoinSymbol, TransactionType.FICR, amount, currentSqlDate));
     }
 
-    public void exchangeCryptoAsset(double amount, String oldCoinSymbol, String newCoinSymbol, Long defaultWalletCurrencyId, Long userId) {
+    public void exchangeCryptoAsset(double amount, String oldCoinSymbol, String newCoinSymbol, String defaultCurrencySymbol, Long userId) {
         CryptoCoin newCoinCrypto = cryptoCoinRepo.findByCryptoCoinSymbol(newCoinSymbol);
-        CryptoCoin oldCoin = cryptoCoinRepo.findByCryptoCoinSymbol(oldCoinSymbol);
-        Long oldCoinId = oldCoin.getCryptoCoinId();
-        Asset oldCoinAsset = assetRepo.findByCoinId(oldCoinId);
+        Asset oldCoinAsset = assetRepo.findByCoinSymbol(oldCoinSymbol);
 
-        FiatCoin defaultCurrency = fiatCoinRepo.findByFiatCoinId(defaultWalletCurrencyId);
-        String defaultCoinSymbol = defaultCurrency.getCoinSymbol();
-        double amountInDefaultCurrency = ExchangeRateMock.exchangeRateService(amount, oldCoinSymbol, defaultWalletCurrencyId);
+        double amountInDefaultCurrency = ExchangeRateMock.exchangeRateService(amount, oldCoinSymbol, defaultCurrencySymbol);
 
         // Set old coin asset - exit, delete or update
         double remainingOldCoinAmount = oldCoinAsset.getCoinAmount() - amount;
@@ -79,47 +72,45 @@ public class AssetService {
             return;
         } else {
             if (remainingOldCoinAmount == 0) {
-                assetRepo.deleteById(oldCoinId);
+                assetRepo.delete(oldCoinAsset);
             } else {
                 oldCoinAsset.setCoinAmount(remainingOldCoinAmount);
             }
         }
 
         // Set intermediary (or final, if newCoin == defaultCurrency) defaultAsset
-        Asset existentDefaultAsset = assetRepo.findByCoinId(defaultWalletCurrencyId);
+        Asset existentDefaultAsset = assetRepo.findByCoinSymbol(defaultCurrencySymbol);
         if (existentDefaultAsset != null) {
             double existingFiatAmount = existentDefaultAsset.getCoinAmount();
             existentDefaultAsset.setCoinAmount(existingFiatAmount + amountInDefaultCurrency);
         } else {
-            Asset newAsset = new Asset(userId, defaultWalletCurrencyId, defaultCoinSymbol, amountInDefaultCurrency);
+            Asset newAsset = new Asset(userId, defaultCurrencySymbol, amountInDefaultCurrency);
             assetRepo.save(newAsset);
         }
 
         // 3 cases of exchange, to : defaultFiatCurrency, new crypto, new fiat
 
-        if (newCoinSymbol.equals(defaultCoinSymbol)) {
+        if (newCoinSymbol.equals(defaultCurrencySymbol)) {
             // CRFI transaction
             // add transaction to DB and exit; defaultAsset already set; oldCrypto already set
             Date currentSqlDate = new Date(System.currentTimeMillis());
-            transactionRepo.save(new Transaction(userId, oldCoinSymbol, defaultCoinSymbol, TransactionType.CRFI, amountInDefaultCurrency, currentSqlDate));
+            transactionRepo.save(new Transaction(userId, oldCoinSymbol, defaultCurrencySymbol, TransactionType.CRFI, amountInDefaultCurrency, currentSqlDate));
             return;
         }
         if (newCoinCrypto != null) {
             // new coin is a crypto
             // CRCR transaction : oldCR-defaultFI , defaultFI-newCR
-            buyCryptoAsset(amountInDefaultCurrency, newCoinSymbol, defaultWalletCurrencyId, userId);
+            buyCryptoAsset(amountInDefaultCurrency, newCoinSymbol, defaultCurrencySymbol, userId);
         } else {
             // new coin is a fiat, but not default one
             // FIFI transaction
-            buyFiatAsset(amountInDefaultCurrency, newCoinSymbol, defaultWalletCurrencyId, userId);
+            buyFiatAsset(amountInDefaultCurrency, newCoinSymbol, defaultCurrencySymbol, userId);
         }
     }
 
-    public void buyFiatAsset(double amountInDefaultCurrency, String newFiatSymbol, long defaultWalletCurrencyId, long userId) {
-        FiatCoin newCoinFiat = fiatCoinRepo.findByFiatCoinSymbol(newFiatSymbol);
-        Long newFiatId = newCoinFiat.getFiatCoinId();
-        Asset existentNewCoinAsset = assetRepo.findByCoinId(newFiatId);
-        Asset defaultCurrencyAsset = assetRepo.findByCoinId(defaultWalletCurrencyId);
+    public void buyFiatAsset(double amountInDefaultCurrency, String newFiatSymbol, String defaultCurrencySymbol, Long userId) {
+        Asset existentNewCoinAsset = assetRepo.findByCoinSymbol(newFiatSymbol);
+        Asset defaultCurrencyAsset = assetRepo.findByCoinSymbol(defaultCurrencySymbol);
         double remainingDefaultAmount;
 
         if (defaultCurrencyAsset != null) {
@@ -139,14 +130,14 @@ public class AssetService {
             assetRepo.delete(defaultCurrencyAsset);
         }
 
-        double amountInNewFiat = ExchangeRateMock.exchangeRateService(amountInDefaultCurrency, newFiatSymbol, defaultWalletCurrencyId);
+        double amountInNewFiat = ExchangeRateMock.exchangeRateService(amountInDefaultCurrency, newFiatSymbol, defaultCurrencySymbol);
         if (existentNewCoinAsset != null) {
             // update asset
             double existentAmount = existentNewCoinAsset.getCoinAmount();
             existentNewCoinAsset.setCoinAmount(existentAmount + amountInNewFiat);
         } else {
             // add asset
-            Asset newCoinFiatAsset = new Asset(userId, newCoinFiat.getFiatCoinId(), newFiatSymbol, amountInNewFiat);
+            Asset newCoinFiatAsset = new Asset(userId, newFiatSymbol, amountInNewFiat);
             assetRepo.save(newCoinFiatAsset);
         }
